renombre de colorPicker, refactor de funcion y agregado prototype en character counter
